backend/requirements.txt
Django==2.1.2
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
django-filter==2.0.0
dj-database-url==0.5.0
psycopg2==2.7.5
PyJWT==1.6.4
pytz==2018.5
waitress==1.1.0
django-pdb==0.6.2
sendgrid==5.6.0
